chore: 장소 상세 모달창, 컴포넌트 분리 (#173)
